* Ssize_t now maps to C# uint, instead of int. * Allocator now takes uints instead of ints. * Some int/uinty CPyMarshal tweaks. * In fact, signedness-fiddling just about everywhere. * Severely cut down schnoz.py -- test runs now look nicer and collect results at end, and we can always reinstate the findfailures stuff if we need it again. * 4 numpy tests currently fail, apparently due to the fact that passing -1 to a function-taking-an-Ssize_t is just fine and dandy in C, but causes an overflow error in C#. I suspect we'll just have to special-case -1, but that's not going in tonight.
by: William

+++ src/Allocator.cs
@@ -5,8 +5,8 @@ namespace Ironclad
 {
     public interface IAllocator
     {
-        IntPtr Alloc(int bytes);
-        IntPtr Realloc(IntPtr old, int bytes);
+        IntPtr Alloc(uint bytes);
+        IntPtr Realloc(IntPtr old, uint bytes);
         bool Contains(IntPtr ptr);
         void Free(IntPtr address);
         void FreeAll();
@@ -17,15 +17,15 @@ namespace Ironclad
         private StupidSet allocated = new StupidSet();
         
         public virtual IntPtr 
-        Alloc(int bytes)
+        Alloc(uint bytes)
         {
-            IntPtr ptr = Marshal.AllocHGlobal(bytes);
+            IntPtr ptr = Marshal.AllocHGlobal((IntPtr)bytes);
             this.allocated.Add(ptr);
             return ptr;
         }
         
         public virtual IntPtr
-        Realloc(IntPtr oldptr, int bytes)
+        Realloc(IntPtr oldptr, uint bytes)
         {
             IntPtr newptr = Marshal.ReAllocHGlobal(oldptr, (IntPtr)bytes);    
             this.allocated.SetRemove(oldptr);
